Response objects available for GephiREST "requests"
GephiREST requests sent in `auto_commit=True` mode will return
`requests.Response` objects.
The Gephi GraphStream API does not specify any description of the
response-to-api-call -- but it might be added later.

[Gephi current;y echoes back the json content of the API request]

The Response objects can be used to diagnose connection problems at best
and used as a confirmation at the very least.